atomic/flags: use typedesc symbol instead of getType
Using `getType` yields an `EnumTy` expression, which is distinct from
the real enum type. This was unnoticed since Nim does not re-sem
`sizeof()` body, allowing the original type to remain.

NimSkull however will re-sem the body of `sizeof()`, causing
crucial pragmas like `{.size: 2.}` to be lost.

Instead of using `getType()`, use the typedesc from generic signature
directly. This is more than enough to get `sizeof()` to produce the
correct size without dealing with the troubles of type-as-ast.

Thanks @zerbina and @saem for help with investigating this issue.