runtimes: handle pthread_cleanup_push/pop implicit scopes
The pthread_cleanup_push/pop pair is implemented as a macro that
generates an implicit C scope. NimSkull is unaware of this and inserts
variable declarations within this implicit scope and cleanups outside of
it, breaking insideout after new C codegen.

This commit scopes the procedure body within a `block` to make sure all
declarations and destructors are kept within pthread_cleanup implicit
scope.